Messaging: add since filter and unread count; handlers and routes; messages table + indexes; wallet/transactions/orders schema updates

+++ internal/marketplace/model.go
@@ -4,21 +4,21 @@ import "time"
 
 // Service represents a service listed by a user
 type Service struct {
-	ID          string    `json:"id"`
-	UserID      string    `json:"user_id"`
-	Title       string    `json:"title"`
-	Description string    `json:"description"`
-	Price       float64   `json:"price"`
-	CreatedAt   time.Time `json:"created_at"`
+    ID          string    `json:"id"`
+    UserID      string    `json:"user_id"`
+    Title       string    `json:"title"`
+    Description string    `json:"description"`
+    Price       int64     `json:"price"`
+    CreatedAt   time.Time `json:"created_at"`
 }
 
 // Order represents an order placed by a user for a service
 type Order struct {
-	ID         string    `json:"id"`
-	ServiceID  string    `json:"service_id"`
-	BuyerID    string    `json:"buyer_id"`
-	SellerID   string    `json:"seller_id"`
-	Amount     float64   `json:"amount"`
-	Status     string    `json:"status"` // pending, completed, cancelled
-	CreatedAt  time.Time `json:"created_at"`
+    ID         string    `json:"id"`
+    ServiceID  string    `json:"service_id"`
+    BuyerID    string    `json:"buyer_id"`
+    SellerID   string    `json:"seller_id"`
+    Amount     int64     `json:"amount"`
+    Status     string    `json:"status"` // pending_acceptance, in_progress, delivered, declined, completed, canceled
+    CreatedAt  time.Time `json:"created_at"`
 }
